address_book/requirements/base.txt
fastapi==0.110.1
uvicorn==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
pyhumps==3.8.0
pydantic-settings==2.2.1